third_party/WebKit/LayoutTests/fast/html/imports/import-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<link id="harness" rel="import" href="resources/import-events-target.html">
</head>
<body>
<script>
description("Test that events can be fired on imported documents.");

function handle(list, id, event)
{
    list.push(event.type + "@" + id);
}

function testEventsFiredOn(doc)
{
    window.caughtFoo = [];
    var root = doc.documentElement;
    var target = doc.getElementById("target");

    target.addEventListener("foo", handle.bind(null, caughtFoo, "target"));
    root.addEventListener("foo", handle.bind(null, caughtFoo, "root"));
    target.dispatchEvent(new CustomEvent("foo", { bubbles: true }));

    shouldBe("window.caughtFoo", "['foo@target', 'foo@root']");
}

testEventsFiredOn(harness.import);
</script>
</body>
</html>
