third_party/WebKit/LayoutTests/fast/html/imports/import-events-expected.txt

Test that events can be fired on imported documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS successfullyParsed is true

TEST COMPLETE
